fix(efb): add required pages for unified efb
