+++ fbw-a32nx/src/systems/instruments/src/EFB/Dashboard/Widgets/FlightWidget.tsx
@@ -10,6 +10,8 @@ import { usePersistentProperty } from '@flybywiresim/fbw-sdk';
 import { toast } from 'react-toastify';
 import { fetchSimbriefDataAction, isSimbriefDataLoaded } from '../../Store/features/simBrief';
 import { useAppSelector, useAppDispatch } from '../../Store/store';
+import { getAirframeType } from 'instruments/src/EFB/Efb';
+import { AC_TYPE } from 'instruments/src/EFB/Enum/Airframe';
 
 import { ScrollableContainer } from '../../UtilComponents/ScrollableContainer';
 import { t } from '../../translation';
@@ -134,7 +136,7 @@ export const FlightWidget = () => {
                         {' '}
                         |
                         {' '}
-                        A320-251N
+                        {AC_TYPE[getAirframeType()]}
                     </h1>
                 )}
             </div>
