+++ fbw-a32nx/src/systems/instruments/src/EFB/Enum/Airframe.ts
@@ -1,4 +1,4 @@
 export enum AC_TYPE {
-    A320_251N = 'A321-251N',
+    A320_251N = 'A320-251N',
     A380_842 = 'A380-842'
 }
